Layout_TestCase/http/tests - 副本/appcache/expected/local-content-expected.txt

CONSOLE MESSAGE: Not allowed to load local resource: stdio.h
Test that documents loaded from application cache don't get access to local myrs accidentally.

Should say SUCCESS:

SUCCESS
